requirements.txt
fastapi
uvicorn
python-dotenv

pydantic

chromadb
sentence-transformers

numpy
pandas

langchain
langchain-text-splitters
langchain-chroma

groq

streamlit

python-multipart